+++ app/src/main/java/com/github/nakawai/newsreader/presentation/sections/SectionListAdapter.kt
@@ -1,4 +1,4 @@
-package com.github.nakawai.newsreader.ui.sections
+package com.github.nakawai.newsreader.presentation.sections
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.github.nakawai.newsreader.databinding.ListItemSectionBinding
 import com.github.nakawai.newsreader.domain.story.Section
-import com.github.nakawai.newsreader.ui.translate
+import com.github.nakawai.newsreader.presentation.translate
 
 class SectionListAdapter(private val listener: OnItemClickListener) : ListAdapter<Section, SectionListAdapter.ViewHolder>(DIFF_CALLBACK) {
