+++ app/src/main/res/layout/activity_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.details.DetailsActivity">
+    tools:context=".presentation.details.DetailsActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
@@ -30,7 +30,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context=".ui.details.DetailsActivity">
+        tools:context=".presentation.details.DetailsActivity">
 
         <TextView
             android:id="@+id/dateView"
