+++ app/src/main/res/layout/activity_sections.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.sections.SectionsActivity">
+    tools:context=".presentation.sections.SectionsActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
@@ -27,7 +27,7 @@
         android:paddingTop="0dp"
         android:paddingBottom="0dp"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context=".ui.articles.ArticlesActivity">
+        tools:context=".presentation.articles.ArticlesActivity">
 
 
         <androidx.recyclerview.widget.RecyclerView
